changed GameRoom to WaitingRoom component and adjusted routing. New Games now routing to waitingroom/<gameID>
