+++ src/components/App/index.js
@@ -9,8 +9,8 @@ import PasswordForgetPage from "../PasswordForget";
 import HomePage from "../Home";
 import Account from "../Account";
 import Admin from "../Admin";
+import WaitingRoom from "../WaitingRoom";
 import GameRoom from "../GameRoom";
-
 import * as ROUTES from "../../constants/routes";
 import { withAuthentication } from "../Session";
 
@@ -28,7 +28,8 @@ const App = () => (
       <Route path={ROUTES.HOME} component={HomePage} />
       <Route path={ROUTES.ACCOUNT} component={Account} />
       <Route path={ROUTES.ADMIN} component={Admin} />
-      <Route path={ROUTES.GAME_ROOM + "/:roomId"} component={GameRoom} />
+      <Route path={ROUTES.GAME_ROOM} component={GameRoom} />
+      <Route path={ROUTES.WAITING_ROOM + "/:roomId"} component={WaitingRoom} />
     </div>
   </Router>
 );
