+++ src/constants/routes.js
@@ -6,3 +6,4 @@ export const ACCOUNT = "/account";
 export const ADMIN = "/admin";
 export const PASSWORD_FORGET = "/pw-forget";
 export const GAME_ROOM = `/gameroom`;
+export const WAITING_ROOM = '/waitingroom'
